refactor: move beforeSend to its own file
